MAINT: stats.linregress: remove out of _stats_mstats_common
